feat: add quote node
