+++ src/components/editor/nodes/quote/styles.module.scss
@@ -1,0 +1,38 @@
+@import "../../../../styles/variables.scss";
+
+.quote {
+  position: relative;
+  border: 1px solid $color-primary;
+  border-radius: 10px;
+  padding: 24px 20px;
+
+  &:before, &:after {
+    content: "";
+    position: absolute;
+
+    background-color: $color-background;
+    width: 26px; height: 18px;
+    padding: 0 4px;
+    color: $color-primary;
+
+    background-image: url("/images/etc/quote-icon.svg");
+    background-size: contain;
+    background-position: center;
+    background-repeat: no-repeat;
+  
+    & > * {
+      fill: $color-primary;
+    }
+  }
+
+  &:before {
+    top: -9px;
+    left: 18px;
+  }
+
+  &:after {
+    bottom: -9px;
+    right: 18px;
+    transform: rotate(180deg);
+  }
+}
