+++ src/components/editor/plugins/placeholder/index.jsx
@@ -1,12 +1,11 @@
 import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
-import { $getRoot, $getSelection, ParagraphNode } from "lexical";
-import { useCallback, useEffect, useState } from "react";
+import { $getRoot, $getSelection } from "lexical";
+import { useEffect, useState } from "react";
 
 import styles from "./styles.module.scss";
 
 const placeholders = {
   paragraph: "'/'를 입력하여 명령어를 사용하세요.",
-  quote: "인용문",
   h2: "제목 1",
   h3: "제목 2",
   h4: "제목 3",
