--- src/components/editor/plugins/quote/index.jsx
@@ -1,6 +1,0 @@
-import { $createQuoteNode } from "@lexical/rich-text";
-import { $createCustomQuoteContainerNode } from "./nodes";
-
-export default function QuotePlugin() {
-  
-}
